HomeController + data to main page(home)

+++ database/migrations/2019_03_22_152332_create_bouquet_of_the_day_table.php
@@ -19,7 +19,7 @@ class CreateBouquetOfTheDayTable extends Migration
             $table->integer('bouquet_id')->unsigned();
             $table->foreign('bouquet_id')->references('id')->on('bouquets');
 
-            $table->integer('discount');
+            $table->decimal('discount',5,3);
         });
     }
